% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1c_kusko_input_prep.R
\name{kusko_prep}
\alias{kusko_prep}
\title{Prepare the Kuskokwim Substock Data for this analysis}
\usage{
kusko_prep(S_dat, H_dat, age_dat)
}
\arguments{
\item{S_dat}{the output file for substock spawner abundance}

\item{H_dat}{the output file for drainage-wide
aggregate harvest summed across all substocks}

\item{age_dat}{the output file for substock-specific age composition data
for substocks that have these data}
}
\description{
Prepare the output data files from the GitHub repository: bstaton1/kusko-multi-sra
into the format needed by the functions in the \code{FitSR} package.
}
