% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_lm_data_prep_function.R
\name{lm_data_prep}
\alias{lm_data_prep}
\title{Prepare data for a linear SRA fit}
\usage{
lm_data_prep(params, obs)
}
\arguments{
\item{params}{a list object created by \code{SimFit::init_sim()} or \code{kusko_data_prep()$params}.
Contains the driving parameters and dimensional variables.}

\item{obs}{a list object created by the functions that generate observations in the
\code{SimFit} package or the \code{kusko_data_prep()$obs} function.}
}
\description{
Takes the output of the \code{SimFit} functions to prepare
for fitting a linear regression model
}
