% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b_tsm_data_prep.R
\name{ssm_data_prep}
\alias{ssm_data_prep}
\title{Prepare data for a state-space SRA fit}
\usage{
ssm_data_prep(params, obs, covariance)
}
\arguments{
\item{params}{a list object created by \code{SimFit::init_sim()} or \code{kusko_data_prep()$params}.
Contains the driving parameters and dimensional variables.}

\item{obs}{a list object created by the functions that generate observations in the
\code{SimFit} package or the \code{kusko_data_prep()$obs} function.}

\item{covariance}{a character vector of length 1: \code{"simple"} or \code{"complex"}}
}
\description{
Takes the output of the \code{SimFit} functions to prepare
for fitting a state-space spawner recruit model
}
