% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3b_ssm_summary.R
\name{ssm_summary}
\alias{ssm_summary}
\title{Quickly summarize the state-space model posteriors}
\usage{
ssm_summary(post, params, model, maturity, diag_plots = F, file = NULL,
  seed = NA, plot_dir)
}
\arguments{
\item{post}{an object of class \code{mcmc.list}}

\item{params}{a list object created by \code{SimFit::init_sim()} or \code{kusko_data_prep()$params}.
Contains the driving parameters and dimensional variables.}

\item{model}{the model identifier: e.g., a number or letter}

\item{maturity}{character vector of length 1: does the model have \code{"simple"} or \code{"complex"}
maturity schedules? The extracted parameters depends on this}

\item{diag_plots}{logical. Do you wish to save traceplots and density plots for convergence diagnostics?
Defaults to \code{FALSE}.}

\item{file}{character vector of length 1: optional file name of the saved diagnostic plots.}

\item{seed}{a numeric vector of length 1: represents a secondary identifier. Typically would be used to identify the seed.
To exclude, set to the default: \code{NA}.}

\item{plot_dir}{character vector representing the directory or full path to the location
where you wish to save the plot files, if \code{diag_plots = TRUE}.
Defaults to \code{NULL}, which will place the files in the working directory.}
}
\description{
Quickly summarize the state-space model posteriors
}
