% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_models.R
\name{write_lme_file}
\alias{write_lme_file}
\title{Write the lm and lme model code}
\usage{
write_lme_file(dir_name = "Model Files", file_name = "lme_model.txt",
  silent = F)
}
\arguments{
\item{dir_name}{the directory name in your working directory to place the file.
If the directory does not exist, it will be created with a warning}

\item{file_name}{the file name to save the file as.}

\item{silent}{logical. Do you wish to suppress the printing of the location of the file?}
}
\description{
Write the lm and lme model code
}
