% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_models.R
\name{write_ssm_1_file}
\alias{write_ssm_1_file}
\title{Write the ssm 1 model code}
\usage{
write_ssm_1_file(dir_name = "Model Files",
  file_name = "ssm_1_model.txt", silent = F)
}
\arguments{
\item{dir_name}{the directory name in your working directory to place the file.
If the directory does not exist, it will be created with a warning}

\item{file_name}{the file name to save the file as.}

\item{silent}{logical. Do you wish to suppress the printing of the location of the file?}
}
\description{
Write the ssm 1 model code
}
