% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ssm_inits.R
\name{gen_ssm_inits}
\alias{gen_ssm_inits}
\title{Generate initial values for a state-space SRA fit}
\usage{
gen_ssm_inits(params, obs, n_chains)
}
\arguments{
\item{params}{a list object created by \code{SimFit::init_sim()} or \code{kusko_data_prep()$params}.
Contains the driving parameters and dimensional variables.}

\item{obs}{a list object created by the functions that generate observations in the
\code{SimFit} package or the \code{kusko_data_prep()$obs} function.}

\item{n_chains}{a numeric vector of length 1: the number of MCMC chains
to generate initial values for}
}
\description{
Takes the output of the \code{SimFit} functions to prepare
for fitting a state-space spawner recruit model using JAGS
}
